migrate more functions
